fintech_hub/requirements.txt
certifi==2022.12.7
tzdata==2022.7
pytz==2022.2.1
setuptools==60.2.0
django-filter==22.1
pip==21.3.1
gunicorn==20.1.0
whitenoise==6.2.0
Pillow==9.2.0
Django==4.1
djangorestframework==3.13.1
asgiref==3.5.2
idna==3.4
requests==2.28.1
psycopg2==2.9.3
charset-normalizer==2.1.1
urllib3==1.26.13
wheel==0.37.1
sqlparse==0.4.2